pkg/utils/hwaddr: migrate code from IPAM pkg
